+++ kipper/core/src/compiler/ast/nodes/declarations/type-declaration/class-declaration/class-member-declaration/class-property-declaration/index.ts
@@ -1,0 +1,3 @@
+export * from "./class-property-declaration";
+export * from "./class-property-declaration-semantics";
+export * from "./class-property-declaration-type-semantics";