+++ kipper/target-ts/src/code-generator.ts
@@ -12,6 +12,8 @@ import type {
 	ParameterDeclaration,
 	TranslatedCodeLine,
 	VariableDeclaration,
+	ClassMethodDeclaration,
+	ClassPropertyDeclaration,
 } from "@kipper/core";
 import { Expression, type LambdaPrimaryExpression } from "@kipper/core";
 import { createTSFunctionSignature, getTSFunctionSignature } from "./tools";
@@ -170,5 +172,38 @@ export class TypeScriptTargetCodeGenerator extends JavaScriptTargetCodeGenerator
 
 		// Return the object primary expression
 		return ["{", "\n", ...indentLines(translatedKeyValuePairs).flat(), "}"];
+	};
+
+	override classMethodDeclaration = async (node: ClassMethodDeclaration): Promise<Array<TranslatedCodeLine>> => {
+		const semanticData = node.getSemanticData();
+		const identifier = semanticData.identifier;
+		const params = semanticData.parameters;
+		const body = semanticData.functionBody;
+		const evaluatedType = TargetTS.getTypeScriptType(semanticData.returnType.getTypeSemanticData().storedType);
+		const returnType = evaluatedType;
+
+		const translatedParams = (
+			await Promise.all(
+				params.map(async (param) => {
+					return await param.translateCtxAndChildren();
+				}),
+			)
+		)
+			.map((param) => [...param.flat(), ", "])
+			.flat();
+		translatedParams.pop(); // Remove the last comma
+
+		return [
+			[identifier, "(", ...translatedParams, ")", ":", " ", returnType],
+			...(await body.translateCtxAndChildren()),
+		];
+	};
+
+	override classPropertyDeclaration = async (node: ClassPropertyDeclaration): Promise<TranslatedCodeLine> => {
+		const semanticData = node.getSemanticData();
+		const identifier = semanticData.identifier;
+		const typeScriptType = TargetTS.getTypeScriptType(semanticData.typeSpecifier.getTypeSemanticData().storedType);
+
+		return [identifier, ":", " ", typeScriptType];
 	};
 }